+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# 4.3.0
+
+New Settings
+
+- Misc > Play Videos on mouse hover
+  - When enabled videos will not auto-play in the Art Select and Token/Tile HUD, and instead will only un-pause when the mouse is hovered over them
+- Misc > Pause Videos on mouse hover out
+  - When enabled videos will pause when the mouse is no longer hovered over them
+
+Compendium Mapper
+
+- With the **Show current images in the 'Art Select' window** enabled the Art Select will now also contain Item descriptions
+- Art Select window will now auto-focus on the search bar and place the cursor at the end of text
+
 # 4.2.1
 
 - Fixing issues with managing scale and mirroring on v10 when setting up custom image configurations